Backend/requirements.txt
Flask
Werkzeug
flask_cors
gunicorn
validate_email
nose
requests
argparse
uuid
flask-swagger-ui
pymongo
email_validator
bcrypt
pyjwt
authlib
virtualenv
itsdangerous==2.0.1
user_agents
fpdf
python-dotenv
